+++ Assets/ScriptableRenderLoop/HDRenderPipeline/Lighting/LightLoop.cs
@@ -1,9 +1,9 @@
 using UnityEngine;
 using UnityEngine.Rendering;
-using UnityEngine.Experimental.Rendering;
+using UnityEngine.Experimental.Rendering.HDPipeline;
 using System;
 
-namespace UnityEngine.Experimental.ScriptableRenderLoop
+namespace UnityEngine.Experimental.Rendering.HDPipeline
 {
     public class BaseLightLoop
     {
