+++ Assets/ScriptableRenderLoop/HDRenderPipeline/Material/Lit/Editor/StandardSpecularToHDLitMaterialUpgrader.cs
@@ -1,8 +1,7 @@
-using System;
 using NUnit.Framework;
 using UnityEngine;
 
-namespace UnityEditor.Experimental.ScriptableRenderLoop
+namespace UnityEditor.Experimental.Rendering.HDPipeline
 {
     class StandardSpecularToHDLitMaterialUpgrader : MaterialUpgrader
     {
