+++ Assets/ScriptableRenderLoop/HDRenderPipeline/Sky/SkyRenderer.cs
@@ -1,10 +1,10 @@
 using UnityEngine.Rendering;
-using UnityEngine.Experimental.Rendering;
+using UnityEngine.Experimental.Rendering.HDPipeline;
 using System.Collections.Generic;
 using System;
 
 
-namespace UnityEngine.Experimental.ScriptableRenderLoop
+namespace UnityEngine.Experimental.Rendering.HDPipeline
 {
     abstract public class SkyRenderer
     {
