+++ Assets/ScriptableRenderLoop/HDRenderPipeline/Utilities.cs
@@ -1,10 +1,10 @@
 using System;
 using UnityEngine.Rendering;
-using UnityEngine.Experimental.Rendering;
+using UnityEngine.Experimental.Rendering.HDPipeline;
 
 using UnityObject = UnityEngine.Object;
 
-namespace UnityEngine.Experimental.ScriptableRenderLoop
+namespace UnityEngine.Experimental.Rendering.HDPipeline
 {
     [Flags]
     public enum ClearFlag
