+++ Assets/ScriptableRenderLoop/RenderPasses/ShadowRenderPass.cs
@@ -4,7 +4,7 @@ using UnityEngine.Profiling;
 using System.Collections.Generic;
 using System;
 
-namespace UnityEngine.Experimental.ScriptableRenderLoop
+namespace UnityEngine.Experimental.Rendering
 {
     [System.Serializable]
     public class ShadowSettings
